refactor: modularize contract definitions and update method naming for clarity
- Moved TRON contract ABI definitions to a dedicated `contracts` module for improved organization.
- Renamed `call_constant` to `trigger_constant_contract` across the codebase for naming consistency.
- Updated `TronChainProvider` methods to use the new `contracts` module.
